+++ src/bin.rs
@@ -186,7 +186,7 @@ fn build_cli() -> clap::Command<'static> {
 #[tokio::main]
 async fn main() {
     if let Err(e) = try_main().await {
-        eprintln!("{}", e);
+        eprintln!("{e}");
         std::process::exit(2);
     }
 }