+++ src/lib/error.rs
@@ -9,7 +9,7 @@ pub enum Error {
     /// Error from the command line parsing (see [clap::Error]).
     Cli(#[from] clap::Error),
     #[error("command failed: {body:?}")]
-    /// Error from a command line process (see [std::process:Command]).
+    /// Error from a command line process (see [std::process::Command]).
     ExitStatus {
         /// Error body.
         body: String,
@@ -67,7 +67,7 @@ pub(crate) fn exit_status_error(exit_status: &ExitStatus) -> Result<()> {
         true => Ok(()),
         false => match exit_status.code() {
             Some(code) => Err(Error::ExitStatus {
-                body: format!("Process terminated with exit code: {}", code),
+                body: format!("Process terminated with exit code: {code}"),
             }),
             None => Err(Error::ExitStatus {
                 body: "Process terminated by signal".to_string(),