+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added issue templates. [#42](https://github.com/jeertmans/languagetool-rust/pull/42)
 - Added dependabot config. [#43](https://github.com/jeertmans/languagetool-rust/pull/43)
 - Added PR template and codecov badge. [#44](https://github.com/jeertmans/languagetool-rust/pull/44)
+- Added missing `#[must_use]`. [#50](https://github.com/jeertmans/languagetool-rust/pull/50)
 
 ### Added